feat: accounting sheets

+++ flow-space-reporting/app/models/metric_value_model.py
@@ -2,5 +2,5 @@ from pydantic import BaseModel
 
 
 class MetricValueModel(BaseModel):
-    value: int | None
-    consumption: int | None
+    value: float | None
+    consumption: float | None
